Extract utility methods into helper class; Better xmldoc and naming
